Use BigDecimal for order prices

+++ backend/src/main/java/com/orderflow/dto/OrderDetailsResponse.java
@@ -4,6 +4,7 @@ import com.orderflow.domain.enums.OrderKind;
 import com.orderflow.domain.enums.OrderStatus;
 import com.orderflow.domain.enums.OrderType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class OrderDetailsResponse {
@@ -11,11 +12,11 @@ public class OrderDetailsResponse {
     private final OrderStatus orderStatus;
     private final OrderType orderType;
     private final OrderKind orderKind;
-    private final Double price;
+    private final BigDecimal price;
     private final Integer quantity;
     private final LocalDateTime createdAt;
 
-    public OrderDetailsResponse(Long orderId, OrderStatus orderStatus, OrderType orderType, OrderKind orderKind, Double price, Integer quantity, LocalDateTime createdAt) {
+    public OrderDetailsResponse(Long orderId, OrderStatus orderStatus, OrderType orderType, OrderKind orderKind, BigDecimal price, Integer quantity, LocalDateTime createdAt) {
         this.orderId = orderId;
         this.orderStatus = orderStatus;
         this.orderType = orderType;
@@ -41,7 +42,7 @@ public class OrderDetailsResponse {
         return orderKind;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
